Fix project references

+++ Source/Meadow.Foundation.Peripherals/Sensors.Radio.Rfid.Rc522/Samples/Sensors.Radio.Rfid.Rc522_Sample/MeadowApp.cs
@@ -52,7 +52,7 @@ namespace MeadowApp
 
         void ShowColorPulse(Color color, int duration = 1000)
         {
-            onboardLed.StartPulse(color, (uint)(duration / 2));
+            onboardLed.StartPulse(color, duration / 2);
             Thread.Sleep(duration);
             onboardLed.Stop();
         }
